+++ lib/opengraph/types.rb
@@ -1,0 +1,12 @@
+module OpenGraph
+  TYPES = {
+    'activity' => %w(activity sport),
+    'business' => %w(bar company cafe hotel restaurant),
+    'group' => %w(cause sports_league sports_team),
+    'organization' => %w(band government non_profit school university),
+    'person' => %w(actor athlete author director musician politician public_figure),
+    'place' => %w(city country landmark state_province),
+    'product' => %w(album book drink food game movie product song tv_show),
+    'website' => %w(blog website)
+  }
+end
